requirements.txt
SQLAlchemy==2.0.41
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3